Made manage devices into a dialog
